requirements.txt
streamlit
pandas
matplotlib
seaborn
Pillow
openpyxl
xlsxwriter
